+++ src/main/java/org/dyn4j/dynamics/joint/PrismaticJoint.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2020 William Bittle  http://www.dyn4j.org/
+ * Copyright (c) 2010-2021 William Bittle  http://www.dyn4j.org/
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are permitted 
@@ -626,8 +626,8 @@ public class PrismaticJoint<T extends PhysicsBody> extends Joint<T> implements S
 	public Vector2 getReactionForce(double invdt) {
 		Vector2 force = new Vector2();
 		// compute the impulse
-		force.x = this.impulse.x * this.perp.x + (this.motorImpulse + this.lowerImpulse + this.upperImpulse) * this.axis.x;
-		force.y = this.impulse.x * this.perp.y + (this.motorImpulse + this.lowerImpulse + this.upperImpulse) * this.axis.y;
+		force.x = this.impulse.x * this.perp.x + (this.motorImpulse + this.lowerImpulse - this.upperImpulse) * this.axis.x;
+		force.y = this.impulse.x * this.perp.y + (this.motorImpulse + this.lowerImpulse - this.upperImpulse) * this.axis.y;
 		// multiply by invdt to obtain the force
 		force.multiply(invdt);
 		return force;