+++ src/main/java/org/dyn4j/dynamics/joint/WheelJoint.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2020 William Bittle  http://www.dyn4j.org/
+ * Copyright (c) 2010-2021 William Bittle  http://www.dyn4j.org/
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are permitted 
@@ -659,8 +659,8 @@ public class WheelJoint<T extends PhysicsBody> extends Joint<T> implements Shift
 	public Vector2 getReactionForce(double invdt) {
 		Vector2 force = new Vector2();
 		// compute the impulse
-		force.x = this.impulse * this.perp.x + (this.springImpulse + this.lowerImpulse + this.upperImpulse) * this.axis.x;
-		force.y = this.impulse * this.perp.y + (this.springImpulse + this.lowerImpulse + this.upperImpulse) * this.axis.y;
+		force.x = this.impulse * this.perp.x + (this.springImpulse + this.lowerImpulse - this.upperImpulse) * this.axis.x;
+		force.y = this.impulse * this.perp.y + (this.springImpulse + this.lowerImpulse - this.upperImpulse) * this.axis.y;
 		// multiply by invdt to obtain the force
 		force.multiply(invdt);
 		return force;