+++ src/test/java/org/dyn4j/collision/continuous/ConservativeAdvancementTest.java
@@ -37,6 +37,9 @@ import junit.framework.TestCase;
 
 /**
  * Test class for the {@link ConservativeAdvancement} class.
+ * <p>
+ * All bodies in this test must be infinite mass so that the
+ * World class doesn't solve their TOI when a step is performed.
  * @author William Bittle
  * @version 4.1.0
  * @since 1.2.0
@@ -88,7 +91,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, this.c2, IDENTITY, new Vector2(), 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, this.c2, IDENTITY, new Vector2(), 0.0, 0.0, 1.0, toi);
 		TestCase.assertTrue(collision);
 		
 		// test the TOI
@@ -120,7 +123,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, this.c2, IDENTITY, new Vector2(), 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, this.c2, IDENTITY, new Vector2(), 0.0, 0.0, 1.0, toi);
 		TestCase.assertTrue(collision);
 		
 		// test the TOI
@@ -153,7 +156,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, 0.0, 1.0, toi);
 		TestCase.assertTrue(collision);
 		
 		// test the TOI
@@ -185,7 +188,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, 0.0, 1.0, toi);
 		TestCase.assertTrue(collision);
 		
 		// test the TOI
@@ -224,7 +227,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, 0.0, 1.0, toi);
 		TestCase.assertTrue(collision);
 		
 		// test the TOI
@@ -260,7 +263,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, 0.0, 1.0, toi);
 		TestCase.assertFalse(collision);
 	}
 	
@@ -285,7 +288,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, 0.0, 1.0, toi);
 		TestCase.assertFalse(collision);
 	}
 	
@@ -314,7 +317,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(this.c1, t1, dp1, 0.0, c2, t2, dp2, 0.0, 0.0, 1.0, toi);
 		TestCase.assertFalse(collision);
 	}
 	
@@ -442,7 +445,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, 0.0, toi);
+		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, 0.0, 0.0, 1.0, toi);
 		TestCase.assertTrue(collision);
 		
 		// get the final transform given the time of impact
@@ -476,7 +479,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, da2, toi);
+		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, da2, 0.0, 1.0, toi);
 		TestCase.assertTrue(collision);
 		
 		// get the final transform given the time of impact
@@ -513,7 +516,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, da2, toi);
+		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, da2, 0.0, 1.0, toi);
 		TestCase.assertFalse(collision);
 	}
 
@@ -536,7 +539,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, da2, toi);
+		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, da2, 0.0, 1.0, toi);
 		TestCase.assertFalse(collision);
 	}
 
@@ -558,7 +561,7 @@ public class ConservativeAdvancementTest {
 		
 		// detect the time of impact
 		TimeOfImpact toi = new TimeOfImpact();
-		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, da2, toi);
+		boolean collision = this.detector.getTimeOfImpact(c1, t1, dp1, da1, c2, t2, dp2, da2, 0.0, 1.0, toi);
 		TestCase.assertFalse(collision);
 	}
 }