+++ src/test/java/org/dyn4j/dynamics/joint/AngleJointTest.java
@@ -32,7 +32,7 @@ import junit.framework.TestCase;
 /**
  * Test case for the {@link AngleJoint} class.
  * @author William Bittle
- * @version 4.0.0
+ * @version 4.0.1
  * @since 2.2.2
  */
 public class AngleJointTest extends AbstractJointTest {