+++ src/test/java/org/dyn4j/geometry/decompose/BayazitTest.java
@@ -35,7 +35,7 @@ import junit.framework.TestCase;
 /**
  * Test case for the {@link Bayazit} class.
  * @author William Bittle
- * @version 3.4.0
+ * @version 4.0.1
  * @since 3.0.0
  */
 public class BayazitTest extends AbstractDecomposeTest {