+++ src/test/java/org/dyn4j/geometry/decompose/EarClippingTest.java
@@ -36,7 +36,7 @@ import junit.framework.TestCase;
 /**
  * Test case for the {@link EarClipping} class.
  * @author William Bittle
- * @version 3.4.0
+ * @version 4.0.1
  * @since 2.2.0
  */
 public class EarClippingTest extends AbstractDecomposeTest {